Same resolution policy as RECON1's 3272c8e9f, re-run on the current tips
(origin/main 903d686e4, origin/port/recon 45bc8223f, merge base 7b2f913fe):

  src/ include/ config/ tools/ notes/ attribution.json .gitignore -> origin/main
  port/                                                           -> ours

This exists only so step 3 can be measured against the tree the landing
merge will actually produce. Nothing here is a landing candidate.

Conflict surface on these tips, which is LARGER than the plan's 161:
  165 modify/delete (10 include/ headers, 155 src/) -- the plan says 7 + 154
   69 content conflicts, of which 44 are src/ or include/ and 27 of those
      carry a real port-side change against the merge base
   38 add/add, of which 36 are src/ or include/ paths both sides created
      independently (none exist at the merge base)

Every one of those 239 src/include paths is a place taking main drops
port-side work, and only the 161 were enumerated before.

+++ src/_ZN11CommonModel13Func_020160ACEj.cpp
@@ -12,7 +12,7 @@ void CommonModel::Func_020160AC(u32 flags)
     for (i = 0; i < n; i++) {
         /* The (long long)(int) launder keeps the ROM's materialized
            address for the read-modify-write. */
-        u32 *f = (u32 *)(((long long)(int)&p->flags));
+        u32 *f = (u32 *)(&p->flags);
         *f |= flags;
         p++;
     }
