+++ src/func_020353f4.c
@@ -1,4 +1,4 @@
 void func_020353f4(char *self)
 {
-    *(unsigned char *)(((long long)(int)(self + 4))) |= 0x40;
+    *(unsigned char *)(self + 4) |= 0x40;
 }
