+++ src/func_02035770.c
@@ -1,4 +1,4 @@
 void func_02035770(char *self)
 {
-    *(unsigned int *)(((long long)(int)(self + 0x10))) &= ~0x200;
+    *(unsigned int *)(self + 0x10) &= ~0x200;
 }
