+++ src/func_02057e00.c
@@ -5,5 +5,5 @@ void func_02057e00(char *self, unsigned char value)
         *(int *)self -= 1;
     }
 
-    *(int *)(((long long)(int)(self + 4))) += 1;
+    *(int *)(self + 4) += 1;
 }
