+++ src/func_0205b4d0.c
@@ -8,7 +8,7 @@ extern struct E7f60 data_020a7f60[];
 
 unsigned char func_0205b4d0(int index, void (*fn)(int), int arg) {
     struct E7f60 *e = &data_020a7f60[index];
-    unsigned char *t = (unsigned char *)(((long long)(int)((char *)e + 8)));
+    unsigned char *t = (unsigned char *)((char *)e + 8);
     data_020a7f60[index].fn = fn;
     e->arg = arg;
     (*t)++;
