+++ src/func_ov002_020bd3a0.c
@@ -5,7 +5,7 @@ extern u8 data_020a0e40;
 extern u16 data_0209f49e;
 extern u16 data_0209f49c;
 
-#define LAUNDER(x) ((int)(((long long)(int)(x))))
+#define LAUNDER(x) ((int)(x))
 
 int func_ov002_020bd3a0(char* p0, u8* p1, unsigned int p2)
 {
