+++ src/func_ov002_020bd4c8.c
@@ -1,5 +1,5 @@
 int func_ov002_020bd4c8(char *self)
 {
-    *(unsigned int *)(((long long)(int)(self + 0xb0))) |= 0x24000000;
+    *(unsigned int *)(self + 0xb0) |= 0x24000000;
     return 1;
 }
