+++ src/func_ov006_020c9aa0.cpp
@@ -42,7 +42,7 @@ extern "C" void func_ov006_020c9aa0(char *c)
     r4v = *(int *)(c + 0x14);
     if (r5v < 0) r5v = -r5v;
     {
-        int *src = (int *)(((long long)(int)(c + 4)));
+        int *src = (int *)(c + 4);
         tmp.x = src[0];
         tmp.y = src[1];
         tmp.z = src[2];
