+++ src/func_ov006_02114fd0.c
@@ -1,5 +1,5 @@
 void func_ov006_02114fd0(char *p)
 {
-    int *a = (int *)(((long long)(int)(p + 0x5984)));
+    int *a = (int *)(p + 0x5984);
     *a = *a + 1;
 }
