+++ src/func_ov080_0212758c.c
@@ -9,7 +9,7 @@ void func_ov080_0212758c(char *a0, char *a1)
 
     t = (*(volatile u16*)(a1 + 0xc) == 0xbf);
     if (t != 0) {
-        a1 = (char*)(int)(((long long)(int)(a1 + 0x5c)));
+        a1 = (char*)(int)(a1 + 0x5c);
         *(int*)(a0 + 0x32c) = *(int*)a1;
         *(int*)(a0 + 0x330) = *(int*)(a1 + 4);
         *(int*)(a0 + 0x334) = *(int*)(a1 + 8);
@@ -25,7 +25,7 @@ void func_ov080_0212758c(char *a0, char *a1)
             *(int*)a1 = v + (int)(((long long)half * sc + 0x800) >> 12);
         }
 
-        a1 = (char*)(int)(((long long)(int)(a1 + 8)));
+        a1 = (char*)(int)(a1 + 8);
         {
             int idx = (*(u16*)(a0 + 0x8e)) >> 4;
             int sc = data_02082214[idx * 2 + 1];
